requirements.txt
torch
torchvision
pandas
numpy
Pillow
matplotlib
scikit-learn
streamlit